fix(carts): Fixes cart modifications not accounting for certain price lists
Price lists may sometimes be bound to certain customer groups. The
addToCartWorkflow and updateLineItemInCartWorkflow workflows were indeed
accounting for the pricing context based on the cart's region, customer
id and currency code, but, unlike in the setPricingContext middleware
for the store/products route, it was not expanding the customer into its
customer groups. For this reason, the pricing module was not calculating
the correct price for the variants.

Fixes #10490